🔨 refactor: refactor: security + UX improvements + module split
Phase 2 fixes:
- SEC-02: zip bomb protection in import
- QUALITY-02: signal handler leak fix
- QUALITY-06: atomic JSON writes (tmp + rename)
- UX-01: validation feedback CSS
- UX-02: save error Banner display
- UX-05: skip auto-detect on template selection
- POLISH-03: geometry save accuracy (non-maximized dimensions)
- POLISH-05: cancel debounce on dialog close
- POLISH-06: context menu action group reuse

High priority:
- 2.6: permission prompt system — camera/mic/geolocation prompt
  via AlertDialog, decisions persisted in permissions.json;
  non-sensitive perms (notifications, clipboard, DRM) auto-granted
- 2.3: import/export on background threads via std::thread::spawn
  + glib::timeout_add_local polling (non-blocking UI)

Medium priority:
- 3.5: collapse Behavior section for new webapps —
  "Advanced Settings..." button reveals on click
- 3.9: split service.rs (735L) into service/ module directory —
  mod.rs (249L), browser.rs (126L), io.rs (130L), migration.rs (255L)

Build: zero warnings (build + clippy), 35/35 tests pass.
